+++ packages/tx-parser/src/stakeOwnership.ts
@@ -1,0 +1,15 @@
+import { Result } from 'neverthrow'
+import { SubStateType, StakeOwnershipT } from './_types'
+import { BufferReaderT } from '@radixdlt/util'
+import { makeBaseStakeSubstateFromBuffer } from './preparedStake'
+
+export const StakeOwnership = {
+	fromBufferReader: (
+		bufferReader: BufferReaderT,
+		lengthData: Buffer,
+	): Result<StakeOwnershipT, Error> =>
+		makeBaseStakeSubstateFromBuffer(SubStateType.STAKE_OWNERSHIP)(
+			bufferReader,
+			lengthData,
+		),
+}