+++ packages/tx-parser/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.2](https://github.com/radixdlt/radixdlt-javascript/compare/@radixdlt/tx-parser@2.1.1...@radixdlt/tx-parser@2.1.2) (2021-07-19)
+
+**Note:** Version bump only for package @radixdlt/tx-parser
+
+
+
+
+
+## [2.1.1](https://github.com/radixdlt/radixdlt-javascript/compare/@radixdlt/tx-parser@2.0.0...@radixdlt/tx-parser@2.1.1) (2021-07-13)
+
+**Note:** Version bump only for package @radixdlt/tx-parser
+
+
+
+
+
 # [2.0.0](https://github.com/radixdlt/radixdlt-javascript/compare/@radixdlt/tx-parser@1.1.31...@radixdlt/tx-parser@2.0.0) (2021-07-13)